Move handle_consensus_transaction
