+++ crates/sui-core/src/authority.rs
@@ -12,7 +12,7 @@ use move_core_types::identifier::Identifier;
 use move_core_types::parser::parse_struct_tag;
 use move_core_types::{language_storage::ModuleId, resolver::ModuleResolver};
 use move_vm_runtime::{move_vm::MoveVM, native_functions::NativeFunctionTable};
-use mysten_metrics::{monitored_scope, spawn_monitored_task};
+use mysten_metrics::spawn_monitored_task;
 use prometheus::{
     register_histogram_with_registry, register_int_counter_with_registry,
     register_int_gauge_with_registry, Histogram, IntCounter, IntGauge, Registry,
@@ -32,7 +32,6 @@ use narwhal_config::{
     Committee as ConsensusCommittee, WorkerCache as ConsensusWorkerCache,
     WorkerId as ConsensusWorkerId,
 };
-use narwhal_types::CommittedSubDag;
 use sui_adapter::{adapter, execution_mode};
 use sui_config::genesis::Genesis;
 use sui_json_rpc_types::{
@@ -75,9 +74,7 @@ use crate::authority::authority_notify_read::NotifyRead;
 use crate::authority::authority_per_epoch_store::AuthorityPerEpochStore;
 use crate::authority_aggregator::TransactionCertifier;
 use crate::checkpoints::CheckpointServiceNotify;
-use crate::consensus_handler::{
-    SequencedConsensusTransaction, VerifiedSequencedConsensusTransaction,
-};
+use crate::consensus_handler::VerifiedSequencedConsensusTransaction;
 use crate::epoch::committee_store::CommitteeStore;
 use crate::epoch::reconfiguration::ReconfigState;
 use crate::execution_driver::execution_process;
@@ -2171,12 +2168,12 @@ impl AuthorityState {
     /// Errors returned by this call are treated as critical errors and cause node to panic.
     pub(crate) async fn handle_consensus_transaction<C: CheckpointServiceNotify>(
         &self,
-        current_epoch: EpochId,
+        epoch_store: &Arc<AuthorityPerEpochStore>,
         transaction: VerifiedSequencedConsensusTransaction,
         checkpoint_service: &Arc<C>,
     ) -> SuiResult {
-        if let Some(certificate) = self
-            .process_consensus_transaction(current_epoch, transaction, checkpoint_service)
+        if let Some(certificate) = epoch_store
+            .process_consensus_transaction(transaction, checkpoint_service, &self.database)
             .await?
         {
             // The certificate has already been inserted into the pending_certificates table by
@@ -2184,112 +2181,6 @@ impl AuthorityState {
             self.transaction_manager.enqueue(vec![certificate]).await?;
         }
         Ok(())
-    }
-
-    /// Depending on the type of the VerifiedSequencedConsensusTransaction wrapper,
-    /// - Verify and initialize the state to execute the certificate.
-    ///   Returns a VerifiedCertificate only if this succeeds.
-    /// - Or update the state for checkpoint or epoch change protocol. Returns None.
-    pub(crate) async fn process_consensus_transaction<C: CheckpointServiceNotify>(
-        &self,
-        current_epoch: EpochId,
-        transaction: VerifiedSequencedConsensusTransaction,
-        checkpoint_service: &Arc<C>,
-    ) -> SuiResult<Option<VerifiedCertificate>> {
-        let _scope = monitored_scope("HandleConsensusTransaction");
-        let VerifiedSequencedConsensusTransaction(SequencedConsensusTransaction {
-            certificate: consensus_output,
-            consensus_index,
-            transaction,
-        }) = transaction;
-        let tracking_id = transaction.get_tracking_id();
-        let epoch_store = match self.load_epoch_store(current_epoch) {
-            Ok(s) => s,
-            Err(err) => {
-                // Epoch has changed while this transaction is being processed. Ignore it.
-                debug!(
-                    "Error loading epoch store in process_consensus_transaction: {:?}",
-                    err
-                );
-                return Ok(None);
-            }
-        };
-        match &transaction.kind {
-            ConsensusTransactionKind::UserTransaction(certificate) => {
-                if certificate.epoch() != current_epoch {
-                    // Epoch has changed after this certificate was sequenced, ignore it.
-                    debug!(
-                        "Certificate epoch ({:?}) doesn't match the current epoch ({:?})",
-                        certificate.epoch(),
-                        current_epoch
-                    );
-                    return Ok(None);
-                }
-                let authority = (&consensus_output.header.author).into();
-                if epoch_store.has_sent_end_of_publish(&authority)? {
-                    // This can not happen with valid authority
-                    // With some edge cases narwhal might sometimes resend previously seen certificate after EndOfPublish
-                    // However this certificate will be filtered out before this line by `consensus_message_processed` call in `verify_consensus_transaction`
-                    // If we see some new certificate here it means authority is byzantine and sent certificate after EndOfPublish (or we have some bug in ConsensusAdapter)
-                    warn!("[Byzantine authority] Authority {:?} sent a new, previously unseen certificate {:?} after it sent EndOfPublish message to consensus", authority.concise(), certificate.digest());
-                    return Ok(None);
-                }
-                // Safe because signatures are verified when VerifiedSequencedConsensusTransaction
-                // is constructed.
-                let certificate = VerifiedCertificate::new_unchecked(*certificate.clone());
-
-                debug!(
-                    ?tracking_id,
-                    tx_digest = ?certificate.digest(),
-                    "handle_consensus_transaction UserTransaction",
-                );
-
-                if !epoch_store
-                    .get_reconfig_state_read_lock_guard()
-                    .should_accept_consensus_certs()
-                {
-                    debug!("Ignoring consensus certificate for transaction {:?} because of end of epoch",
-                    certificate.digest());
-                    return Ok(None);
-                }
-
-                if certificate.contains_shared_object() {
-                    epoch_store
-                        .record_shared_object_cert_from_consensus(
-                            &transaction,
-                            &certificate,
-                            consensus_index,
-                            &self.database,
-                        )
-                        .await?;
-                } else {
-                    epoch_store
-                        .record_owned_object_cert_from_consensus(
-                            &transaction,
-                            &certificate,
-                            consensus_index,
-                        )
-                        .await?;
-                }
-
-                Ok(Some(certificate))
-            }
-            ConsensusTransactionKind::CheckpointSignature(info) => {
-                checkpoint_service.notify_checkpoint_signature(info)?;
-                epoch_store
-                    .record_consensus_transaction_processed(&transaction, consensus_index)?;
-                Ok(None)
-            }
-            ConsensusTransactionKind::EndOfPublish(authority) => {
-                debug!("Received EndOfPublish from {:?}", authority.concise());
-                epoch_store.record_end_of_publish(
-                    *authority,
-                    transaction.key(),
-                    consensus_index,
-                )?;
-                Ok(None)
-            }
-        }
     }
 
     pub async fn create_advance_epoch_tx_cert(
