+++ crates/sui-core/src/authority/authority_per_epoch_store.rs
@@ -899,6 +899,94 @@ impl AuthorityPerEpochStore {
         Ok(VerifiedSequencedConsensusTransaction(transaction))
     }
 
+    /// Depending on the type of the VerifiedSequencedConsensusTransaction wrapper,
+    /// - Verify and initialize the state to execute the certificate.
+    ///   Returns a VerifiedCertificate only if this succeeds.
+    /// - Or update the state for checkpoint or epoch change protocol. Returns None.
+    pub(crate) async fn process_consensus_transaction<C: CheckpointServiceNotify>(
+        &self,
+        transaction: VerifiedSequencedConsensusTransaction,
+        checkpoint_service: &Arc<C>,
+        parent_sync_store: impl ParentSync,
+    ) -> SuiResult<Option<VerifiedCertificate>> {
+        let _scope = monitored_scope("HandleConsensusTransaction");
+        let VerifiedSequencedConsensusTransaction(SequencedConsensusTransaction {
+            certificate: consensus_output,
+            consensus_index,
+            transaction,
+        }) = transaction;
+        let tracking_id = transaction.get_tracking_id();
+        match &transaction.kind {
+            ConsensusTransactionKind::UserTransaction(certificate) => {
+                if certificate.epoch() != self.epoch() {
+                    // Epoch has changed after this certificate was sequenced, ignore it.
+                    debug!(
+                        "Certificate epoch ({:?}) doesn't match the current epoch ({:?})",
+                        certificate.epoch(),
+                        self.epoch()
+                    );
+                    return Ok(None);
+                }
+                let authority = (&consensus_output.header.author).into();
+                if self.has_sent_end_of_publish(&authority)? {
+                    // This can not happen with valid authority
+                    // With some edge cases narwhal might sometimes resend previously seen certificate after EndOfPublish
+                    // However this certificate will be filtered out before this line by `consensus_message_processed` call in `verify_consensus_transaction`
+                    // If we see some new certificate here it means authority is byzantine and sent certificate after EndOfPublish (or we have some bug in ConsensusAdapter)
+                    warn!("[Byzantine authority] Authority {:?} sent a new, previously unseen certificate {:?} after it sent EndOfPublish message to consensus", authority.concise(), certificate.digest());
+                    return Ok(None);
+                }
+                // Safe because signatures are verified when VerifiedSequencedConsensusTransaction
+                // is constructed.
+                let certificate = VerifiedCertificate::new_unchecked(*certificate.clone());
+
+                debug!(
+                    ?tracking_id,
+                    tx_digest = ?certificate.digest(),
+                    "handle_consensus_transaction UserTransaction",
+                );
+
+                if !self
+                    .get_reconfig_state_read_lock_guard()
+                    .should_accept_consensus_certs()
+                {
+                    debug!("Ignoring consensus certificate for transaction {:?} because of end of epoch",
+                    certificate.digest());
+                    return Ok(None);
+                }
+
+                if certificate.contains_shared_object() {
+                    self.record_shared_object_cert_from_consensus(
+                        &transaction,
+                        &certificate,
+                        consensus_index,
+                        parent_sync_store,
+                    )
+                    .await?;
+                } else {
+                    self.record_owned_object_cert_from_consensus(
+                        &transaction,
+                        &certificate,
+                        consensus_index,
+                    )
+                    .await?;
+                }
+
+                Ok(Some(certificate))
+            }
+            ConsensusTransactionKind::CheckpointSignature(info) => {
+                checkpoint_service.notify_checkpoint_signature(info)?;
+                self.record_consensus_transaction_processed(&transaction, consensus_index)?;
+                Ok(None)
+            }
+            ConsensusTransactionKind::EndOfPublish(authority) => {
+                debug!("Received EndOfPublish from {:?}", authority.concise());
+                self.record_end_of_publish(*authority, transaction.key(), consensus_index)?;
+                Ok(None)
+            }
+        }
+    }
+
     pub fn handle_commit_boundary<C: CheckpointServiceNotify>(
         &self,
         committed_dag: &Arc<CommittedSubDag>,
